Changed backup method so Zed would update cleanly during demo
CHanged /Bob/'s script to do a copy to "PREVIOUS" instead of a rename so the changes would show up instantly in Zed.

Also /Bob/ keeps alternating ways to access private "elements" (that should already be readily available but most developers miss that. So too apparently /Bob/, who is only as experienced as his human peers.