Added defensive test against Jetlagged Intern
But /Bob/'s implementation was already too specific. Perhaps in order to trick Bob into writing a too-simple "solution" (i.e., using array sizes instead of a complete implementation) we need to change the domain from set arithmetic to "tracking tagged endangered animals"? But then...is the domain too weird for a demo?