Capture BOBs_FINSET_myset.rb
Just for fun. I liked how /Bob/ "chose" to use algebraic operators rather than iterating over set elements. Shows off the power of Ruby. THis is one of the rare times when the solution used operators exclusively: &, |, and -

It is a *clever* design, which I would normally argue *against* unless everyone on the team understood those operators in this context.